Inline WebSocket.hs into Runtime.hs
